requirements.txt
praw
google-cloud-storage
google-cloud-secret-manager